requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.1.0
boto3==1.40.28
botocore==1.40.28
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
django-anymail==12.0
django-filter==25.1
django-graphene-auth==1.1.1
django-graphql-jwt==0.4.0
django-graphql-ratelimit==1.0.2
django-ipware==7.0.1
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-redis==6.0.0
ffmpeg-python==0.2.0
future==1.0.0
graphene==3.4.3
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.2
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.13
pillow==11.3.0
platformdirs==4.4.0
promise==2.3
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
s3transfer==0.14.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
zope.interface==7.2
faker==24.0.0
